Add AuthExample
