+++ DevPods/Authentication/AuthDev/AuthExample/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  AuthExample
+//
+//  Created by a.gachkovskaya on 29.02.2020.
+//  Copyright © 2020 a.gachkovskaya. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
