--- DevPods/Authentication/Example/.swiftlint.yml
@@ -1,63 +1,0 @@
-identifier_name:
-  excluded:
-    - id
-    - vm
-    - to
-nesting:
-  type_level: 3
-line_length:
-  warning: 150
-  error: 200
-excluded:
-  - Pods
-  - R.generated.swift
-opt_in_rules:
-  - anyobject_protocol
-  - closure_end_indentation
-  - closure_spacing
-  - collection_alignment
-  - contains_over_filter_count
-  - contains_over_filter_is_empty
-  - contains_over_first_not_nil
-  - contains_over_range_nil_comparison
-  - discouraged_object_literal
-  - empty_collection_literal
-  - empty_count
-  - empty_string
-  - first_where
-  - flatmap_over_map_reduce
-  - identical_operands
-  - last_where
-  - legacy_multiple
-  - legacy_random
-  - multiline_arguments
-  - multiline_arguments_brackets
-  - multiline_function_chains
-  - multiline_literal_brackets
-  - multiline_parameters
-  - multiline_parameters_brackets
-  - operator_usage_whitespace
-  - overridden_super_call
-  - override_in_extension
-  - pattern_matching_keywords
-  - private_action
-  - private_outlet
-  - prohibited_super_call
-  - quick_discouraged_focused_test
-  - quick_discouraged_pending_test
-  - raw_value_for_camel_cased_codable_enum
-  - reduce_into
-  - redundant_nil_coalescing
-  - redundant_type_annotation
-  - sorted_first_last
-  - strong_iboutlet
-  - switch_case_on_newline
-  - toggle_bool
-  - unavailable_function
-  - unneeded_parentheses_in_closure_argument
-  - unused_declaration
-  - unused_import
-  - vertical_parameter_alignment_on_call
-  - vertical_whitespace_closing_braces
-  - vertical_whitespace_opening_braces
-  - yoda_condition
